requirements.txt
twitchio>=2.9.1
pytchat>=0.5.5
Flask>=3.0.3
Flask-SocketIO>=5.3.6
eventlet>=0.36.0
PyYAML>=6.0
requests>=2.32.3
psutil>=5.9.0